backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
google-generativeai==0.8.0
pymupdf==1.23.26
python-docx==1.1.0
pydantic>=2.5.3
pydantic-settings>=2.4.0
sqlalchemy==2.0.25
python-dotenv==1.0.0
reportlab==4.0.9
qdrant-client==1.7.3
sentence-transformers==2.3.1
numpy==1.26.3
scikit-learn==1.3.2

# LangChain dependencies for improved AI pipeline
langchain-core>=0.3.32
langchain-google-genai>=2.0.8
langchain-community>=0.3.16
